feat: publish 'VideoFileReplaced' message on video file change

+++ src/video/video.broker.ts
@@ -12,7 +12,7 @@ export class VideoBroker {
             'uploader.video.upload.succeeded',
             async (data: any) => {
                 await this.updateAfterUpload(data);
-                VideoBroker.publishVideoUploaded(data.id, data.key);
+                VideoBroker.publishVideoUploaded(data.id, data.key, data.userId);
             });
 
         rabbit.subscribe(
@@ -43,12 +43,12 @@ export class VideoBroker {
             });
     }
 
-    public static publishVideoUploaded(id: string, key: string) {
+    public static publishVideoUploaded(id: string, key: string, userId: string) {
         rabbit.publish(
             'application',
             'topic',
             'videoService.video.upload.succeeded',
-            { id, key },
+            { id, key, userId },
             { persistent: true },
         );
     }
@@ -69,6 +69,22 @@ export class VideoBroker {
         );
     }
 
+    public static publishVideoFileReplaced(id: string, userId: string, contentPath: string, previewPath: string, thumbnailPath: string) {
+        rabbit.publish(
+            'application',
+            'topic',
+            'videoService.video.replaced.succeeded',
+            {
+                id,
+                userId,
+                contentPath,
+                previewPath,
+                thumbnailPath,
+            },
+            { persistent: true },
+        );
+    }
+
     private static updateAfterUpload(data: { id: string, key: string }) {
         return VideoManager.updateById(data.id, {
             status: VideoStatus.UPLOADED,
@@ -76,7 +92,14 @@ export class VideoBroker {
         });
     }
 
-    private static deleteAfterCancellation(data: { id: string }) {
+    private static async deleteAfterCancellation(data: { id: string }) {
+        const isPublished = await VideoManager.isVideoPublished(data.id);
+
+        // Relevent for reuploades
+        if (isPublished) {
+            return VideoManager.updateById(data.id, { status: VideoStatus.READY });
+        }
+
         return VideoManager.deleteById(data.id);
     }
 
@@ -85,13 +108,15 @@ export class VideoBroker {
         contentPath: string,
         thumbnailPath: string,
         previewPath: string,
+        userId: string,
     }) {
         return VideoManager.updateById(data.id, {
             contentPath: data.contentPath,
             thumbnailPath: data.thumbnailPath,
             previewPath: data.previewPath,
             status: VideoStatus.READY,
-        });
+        },
+            data.userId);
     }
 
     public static updateStatusFailed(data: { id: string }) {
